+++ ui/src/data_explorer/constants/index.ts
@@ -298,7 +298,7 @@ export const METAQUERY_TEMPLATE_OPTIONS: Array<
 ]
 
 export const WRITE_DATA_DOCS_LINK =
-  'https://docs.influxdata.com/influxdb/latest/write_protocols/line_protocol_tutorial/'
+  'https://docs.influxdata.com/influxdb/v1.8/write_protocols/line_protocol_tutorial/'
 
 export const DEFAULT_TIME_RANGE = timeRanges.find(
   tr => tr.lower === 'now() - 1h'